Updated termsandjargon.md
Changed the name of DailySnapshot to HourlySnapshot
Changed the name of DailyIncremantal to HourlyIncemental
Changed the name of DailyDifferntial to HourlyDifferential

With the repercussions that we must drop Weekly and MontlySnapshot,
and just use full to avoid redundancy.

Hindsightly TODAYS_BACKUP_FOLDER IS A TERRIBLE NAME.
BACKUP_INTERVAL_FOLDER sounds better.